backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.3
gunicorn==21.2.0

